@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using ToDo.Domain.Migrations;
 using ToDo.Extensibility.Dto;
 
@@ -21,12 +21,12 @@ namespace ToDo.UI
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddToDo();
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddMvc();
             services.Configure<ConfigurationSettings>(Configuration.GetSection("ConfigurationSettings"));
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -38,13 +38,13 @@ namespace ToDo.UI
             }
 
             app.UseStaticFiles();
-            app.UseCookiePolicy();
 
-            app.UseMvc(routes =>
+            app.UseRouting();
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=ToDo}/{action=Index}/{id?}");
+                    pattern: "{controller=ToDo}/{action=Index}/{id?}");
             });
 
             app.UpdateSchema();
